[codegen/python] Support generating pulumiplugin.json
This is a new optional file that can be generated and included in a python package that provides information about a Pulumi plugin (resource, name, version, server) associated with the package.

When this file isn't present, the plugin name and version are determined from the package name and version (if possible). If the package is an alpha/beta/rc/dev release, then the plugin version cannot be determined without it being provided in this file.

Also, the server cannot be determined without this file. (If not present, the default server is used.)